analytics-engine/pyproject.toml
[tool.poetry]
name = "analytics-engine"
version = "0.1.0"
description = ""
authors = ["han-so1omon <esolomon1221@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.1"
torch = "^2.4.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
